refactor(top10): extract ranking-surface into standalone component (AGN-1390)
Move ranking-board subtree from Top10Page.tsx (lines 424-1130) into a new
sibling module src/components/top10/ranking-surface.tsx per parent AGN-936.

Moved: RankingPanel, FilterRow, SUPPORTED_METRICS, ChipLabel, Chip,
EmptyRows, RankingMetaStrip, plus their internal helpers (PanelHead,
CornerDots, RankRow, Avatar, Body, Badge, Metric, Delta) and windowLabel.

Top10Page keeps page-shell concerns: URL/state sync, route param parsing,
PageHead, category tabs, layout wiring, share surface, more-grid surfaces.
CornerDots and windowLabel are re-imported where Top10Page still needs them
(share-stack head, more-grid mini, tweet text).

Net: Top10Page 1741 -> 1030 LOC (-711); ranking-surface 632 LOC.
No behavior change. typecheck clean.
